# docs/refactor_reports/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nuzum API v2",
    "version": "2.0.0",
    "description": "Auto-generated baseline Swagger for v2 endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v2/attendance/dashboard": {
      "get": {
        "operationId": "dashboard",
        "summary": "dashboard",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/employee-requests/requests/invoice": {
      "post": {
        "operationId": "create_invoice",
        "summary": "create invoice",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}